backend/pyproject.toml
[project]
name = "meblog-backend"
version = "1.0.0"
description = "Meblog Backend API"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.110.0",
    "uvicorn[standard]==0.27.0",
    "sqlalchemy==2.0.25",
    "pydantic==2.6.0",
    "pydantic-settings==2.1.0",
    "python-jose[cryptography]==3.3.0",
    "bcrypt>=4.1.0",
    "python-multipart==0.0.6",
    "alembic==1.13.1",
    "python-slugify==8.0.1",
    "email-validator>=2.1.0",
    "loguru>=0.7.3",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.31.0",
    "aiosqlite>=0.22.1",
]

[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
